backend/requirements.txt
fastapi
uvicorn[standard]
requests
beautifulsoup4
openai>=1.0.0
google-genai>=1.0.0
anthropic>=0.25.0
python-multipart
pydantic>=2.0.0
urllib3
